+++ plutus-metatheory/src/MAlonzo/Code/Cost/Model.hs
@@ -827,13 +827,13 @@ d_getModel_432 v0 v1
                                (\ v6 ->
                                   coe
                                     MAlonzo.Code.Data.String.Properties.du_'8776''45'reflexive_8
-                                    (coe MAlonzo.Code.Builtin.d_showBuiltin_428 (coe v0)))
+                                    (coe MAlonzo.Code.Builtin.d_showBuiltin_432 (coe v0)))
                                (coe
                                   MAlonzo.Code.Data.List.Relation.Binary.Pointwise.Properties.du_decidable_112
                                   (coe MAlonzo.Code.Data.Char.Properties.d__'8799'__14)
                                   (coe
                                      MAlonzo.Code.Agda.Builtin.String.d_primStringToList_12
-                                     (MAlonzo.Code.Builtin.d_showBuiltin_428 (coe v0)))
+                                     (MAlonzo.Code.Builtin.d_showBuiltin_432 (coe v0)))
                                   (coe
                                      MAlonzo.Code.Agda.Builtin.String.d_primStringToList_12 v4))) in
                   coe
@@ -853,7 +853,7 @@ d_getModel_432 v0 v1
                                           MAlonzo.Code.Data.List.NonEmpty.Base.d_tail_32
                                           (coe
                                              MAlonzo.Code.Builtin.Signature.d_args_86
-                                             (coe MAlonzo.Code.Builtin.d_signature_298 (coe v0))))))
+                                             (coe MAlonzo.Code.Builtin.d_signature_302 (coe v0))))))
                                  (coe v5))
                        else coe d_getModel_432 (coe v0) (coe v3))
              _ -> MAlonzo.RTE.mazUnreachableError
@@ -878,7 +878,7 @@ d_lookupModel_474 v0 v1
         -> case coe v2 of
              MAlonzo.Code.Agda.Builtin.Sigma.C__'44'__32 v4 v5
                -> let v6
-                        = MAlonzo.Code.Builtin.d_decBuiltin_426 (coe v4) (coe v1) in
+                        = MAlonzo.Code.Builtin.d_decBuiltin_430 (coe v4) (coe v1) in
                   coe
                     (case coe v6 of
                        MAlonzo.Code.Relation.Nullary.Decidable.Core.C__because__32 v7 v8
@@ -929,4 +929,4 @@ d_createMap_536 v0
          (coe
             MAlonzo.Code.Data.List.Base.du_map_22
             (coe (\ v1 -> d_getModel_432 (coe v1) (coe v0)))
-            (coe MAlonzo.Code.Builtin.d_builtinList_430)))
+            (coe MAlonzo.Code.Builtin.d_builtinList_434)))