+++ plutus-metatheory/src/MAlonzo/Code/Untyped.hs
@@ -571,7 +571,7 @@ d_decUTm_336 v0 v1
                 MAlonzo.Code.RawU.C_UBuiltin_220 v4
                   -> coe
                        MAlonzo.Code.Relation.Nullary.Decidable.Core.d_does_28
-                       (coe MAlonzo.Code.Builtin.d_decBuiltin_426 (coe v3) (coe v4))
+                       (coe MAlonzo.Code.Builtin.d_decBuiltin_430 (coe v3) (coe v4))
                 _ -> coe v2
          MAlonzo.Code.RawU.C_UDelay_222 v3
            -> case coe v1 of